resume-shortlister-app/backend/requirements.txt
flask
torch
transformers
scikit-learn
jobl